pyproject.toml
[project]
name = "filmssystem"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikita Orlovsky",email = "orlovski15555@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.120.1,<0.121.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aioboto3 (>=15.5.0,<16.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
